requirements.txt
django == 4.2.7
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.3.0
django-cors-headers == 4.3.0
django-crispy-forms == 2.1
pyscopg2 == 66.0.2
crispy-bootstrap4 == 2023.1
importlib-resources==5.12
Pillow==10.1.0
gunicorn==21.2.0
psycopg2==2.9.9
asgiref==3.7.2
pytz==2023.3.post1
sqlparse==0.4.4
django-allauth==0.60.0
python-decouple==3.8
django-cors-headers==4.3.0
httpx==0.26.0
daphne==4.1.0
celery==5.3.6
django-q==1.3.9
django-http-exceptions==1.4.0
drf-spectacular==0.27.1
overpy==0.7
pyotp==2.9.0
twilio==9.0.3
python-dotenv==1.0.0
coverage==7.6.4
pytest==8.3.1
pytest-cov==4.1.0   
pytest-django==4.5.2
